implement mmap dynamic growth + cleanup.
This commit implements dynamic mmap growth. Upon an MDB_MAP_FULL error, an
mmap expansion will be attempted.

The sizing is determined by the MmapGrowthStepFactor and MmapGrowthStepMax.

* MmapGrowthStepFactor => determines the next map size when a write fails.
  The current size is multiplied by the factor, and rounded up to the next
  value divisible by os.Getpagesize(), to obtain the new map size, which is
  applied with mdb_env_set_mapsize.
* MmapGrowthStepMax => is the maximum step size by which we'll grow the mmap.

LMDB requires no transactions to be in progress when a resize takes place.
LMDB will fail if write txns are in progress, but not if there are only read txns.
Therefore, we can't rely on LMDB to correctly reject the expansion, and instead
we need to synchronize. Otherwise we risk changing the virtual address space
and keeping hanging pointers which cause segmentation faults.

Cursors are a problem, as long-running cursors can prevent expansions from
taking place entirely. Therefore, we interrupt running cursors, so that
they yield temporarily for the mmap expansion to take place. The cursor
is then resumed from the position where it was interrupted.

This commit also migrates to github.com/ledgerwatch/lmdb-go.

It also performs some cleanup and godocs enhancement.